Переместил типы в main-page
