+++ project/src/components/pages/main-page/constants/constants.ts
@@ -1,5 +1,4 @@
-import { LocationTabName } from 'src/types/main-page-location-tabs';
-import { MainPageSortOfferType } from 'src/types/main-page';
+import { LocationTabName, MainPageSortOfferType } from 'src/types/main-page';
 
 export const DEFAULT_LOCATION_TAB_NAME: LocationTabName = LocationTabName.PARIS;
 export const DEFAULT_SORTING_OFFER_TYPE = MainPageSortOfferType.Popular;
