+++ project/src/components/pages/main-page/main-page-location-tabs.tsx
@@ -4,11 +4,11 @@ import { Link, useSearchParams } from 'react-router-dom';
 import Tabs from 'src/components/shared/tabs/tabs';
 import { buildLocationTabs } from 'src/components/pages/main-page/helpers/build-location-tabs';
 import { parseSearchParams } from 'src/helpers/parse-search-params';
-import { LocationTabSearchParam } from 'src/types/main-page-location-tabs';
 import { Tab } from 'src/types/tabs';
 import { DEFAULT_LOCATION_TAB_NAME } from 'src/components/pages/main-page/constants/constants';
 import { useAppDispatch, useAppSelector } from 'src/hooks';
 import { setCurrentCity } from 'src/store/actions/actions';
+import { LocationTabSearchParam } from 'src/types/main-page';
 
 const locationTabs: Tab[] = buildLocationTabs();
