--- project/src/types/main-page-location-tabs.ts
@@ -1,12 +1,0 @@
-export type LocationTabSearchParam = {
-  country: LocationTabName;
-};
-
-export enum LocationTabName {
-  PARIS = 'Paris',
-  COLOGNE = 'Cologne',
-  BRUSSELS = 'Brussels',
-  AMSTERDAM = 'Amsterdam',
-  HAMBURG = 'Hamburg',
-  DUSSELDORF = 'Dusseldorf',
-}
