+++ project/src/types/main-page.ts
@@ -16,3 +16,16 @@ export enum MainPageSortOfferType {
   PriceUp = 'Price: high to low',
   RatedUp = 'Top rated first',
 }
+
+export type LocationTabSearchParam = {
+  country: LocationTabName;
+};
+
+export enum LocationTabName {
+  PARIS = 'Paris',
+  COLOGNE = 'Cologne',
+  BRUSSELS = 'Brussels',
+  AMSTERDAM = 'Amsterdam',
+  HAMBURG = 'Hamburg',
+  DUSSELDORF = 'Dusseldorf',
+}
